+++ .gitignore
@@ -476,3 +476,4 @@ $RECYCLE.BIN/
 
 
 # End of https://www.gitignore.io/api/git,linux,csharp,windows,visualstudio
+Test
