+++ Raylib-cs.sln
@@ -9,9 +9,17 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Generator", "Generator\Gene
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples", "Examples\Examples.csproj", "{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test.NetCore", "Test.NetCore\Test.NetCore.csproj", "{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test.NetFX", "Test.NetFX\Test.NetFX.csproj", "{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Test.Common", "Test.Common\Test.Common.shproj", "{36E54E2A-5899-4D4D-9D78-120DB1B4C7B2}"
 EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Test.Common\Test.Common.projitems*{36e54e2a-5899-4d4d-9d78-120db1b4c7b2}*SharedItemsImports = 13
+		Test.Common\Test.Common.projitems*{55d7bb37-3dc7-4a33-b21e-4f549ec988b5}*SharedItemsImports = 4
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
@@ -23,52 +31,60 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{9F30944C-415B-4763-91C7-81721117879D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9F30944C-415B-4763-91C7-81721117879D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x64.Build.0 = Debug|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x86.Build.0 = Debug|Any CPU
+		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x64.ActiveCfg = Debug|x64
+		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x64.Build.0 = Debug|x64
+		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x86.ActiveCfg = Debug|x86
+		{9F30944C-415B-4763-91C7-81721117879D}.Debug|x86.Build.0 = Debug|x86
 		{9F30944C-415B-4763-91C7-81721117879D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9F30944C-415B-4763-91C7-81721117879D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Release|x64.ActiveCfg = Release|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Release|x64.Build.0 = Release|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Release|x86.ActiveCfg = Release|Any CPU
-		{9F30944C-415B-4763-91C7-81721117879D}.Release|x86.Build.0 = Release|Any CPU
+		{9F30944C-415B-4763-91C7-81721117879D}.Release|x64.ActiveCfg = Release|x64
+		{9F30944C-415B-4763-91C7-81721117879D}.Release|x64.Build.0 = Release|x64
+		{9F30944C-415B-4763-91C7-81721117879D}.Release|x86.ActiveCfg = Release|x86
+		{9F30944C-415B-4763-91C7-81721117879D}.Release|x86.Build.0 = Release|x86
 		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x64.Build.0 = Debug|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x86.Build.0 = Debug|Any CPU
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x64.ActiveCfg = Debug|x64
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x64.Build.0 = Debug|x64
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x86.ActiveCfg = Debug|x86
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Debug|x86.Build.0 = Debug|x86
 		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x64.ActiveCfg = Release|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x64.Build.0 = Release|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x86.ActiveCfg = Release|Any CPU
-		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x86.Build.0 = Release|Any CPU
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x64.ActiveCfg = Release|x64
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x64.Build.0 = Release|x64
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x86.ActiveCfg = Release|x86
+		{063F21F1-12D3-41C6-B598-125C725955B1}.Release|x86.Build.0 = Release|x86
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Debug|x64.ActiveCfg = Debug|x64
-		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Debug|x64.Build.0 = Debug|x64
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Debug|x86.ActiveCfg = Debug|x86
-		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Debug|x86.Build.0 = Debug|x86
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Release|x64.ActiveCfg = Release|x64
-		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Release|x64.Build.0 = Release|x64
 		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Release|x86.ActiveCfg = Release|x86
-		{EF58DEDB-3ACA-46F5-A443-1B93D28F206C}.Release|x86.Build.0 = Release|x86
 		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x64.Build.0 = Debug|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x86.Build.0 = Debug|Any CPU
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x64.ActiveCfg = Debug|x64
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x64.Build.0 = Debug|x64
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x86.ActiveCfg = Debug|x86
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Debug|x86.Build.0 = Debug|x86
 		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x64.ActiveCfg = Release|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x64.Build.0 = Release|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x86.ActiveCfg = Release|Any CPU
-		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x86.Build.0 = Release|Any CPU
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x64.ActiveCfg = Release|x64
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x64.Build.0 = Release|x64
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x86.ActiveCfg = Release|x86
+		{1298EE9D-E480-4014-BED0-C7CFE94C4AD7}.Release|x86.Build.0 = Release|x86
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Debug|x64.ActiveCfg = Debug|x64
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Debug|x64.Build.0 = Debug|x64
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Debug|x86.ActiveCfg = Debug|x86
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Debug|x86.Build.0 = Debug|x86
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Release|x64.ActiveCfg = Release|x64
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Release|x64.Build.0 = Release|x64
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Release|x86.ActiveCfg = Release|x86
+		{55D7BB37-3DC7-4A33-B21E-4F549EC988B5}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
