+++ Test.NetCore/Properties/launchSettings.json
@@ -1,0 +1,7 @@
+{
+  "profiles": {
+    "Test.NetCore": {
+      "commandName": "Project"
+    }
+  }
+}
