+++ Test.NetFX/Program.cs
@@ -1,0 +1,12 @@
+﻿using Test.Common;
+
+namespace Test.NetFX
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Common.Test.Run();
+        }
+    }
+}
